Changed thread ID to use 32 bit identifiers rather than excessive 64 bit

+++ Osiris/Libraries/Ausar/Threading/Threading.cpp
@@ -5,7 +5,7 @@ namespace Ausar
 {
     namespace Threading
     {
-        uint64_t ThreadTable::GetThreadIDByName(const char * Name) const
+        uint32_t ThreadTable::GetThreadIDByName(const char * Name) const
         {
             for( size_t i = 0; i < 64; i++ )
             {
@@ -13,7 +13,7 @@ namespace Ausar
                 {
                     if( std::strcmp(Entries[i].ThreadName, Name) == 0 )
                     {
-                        return Entries[i].ThreadID;
+                        return static_cast<uint32_t>(Entries[i].ThreadID);
                     }
                 }
             }
